refactor(backend): 重构 ErrorHandler 模块到错误处理目录
- 为什么改：ErrorHandler.ts 是函数式工具集合，不符合 services/ 目录的服务类模式；与 MCPErrors.ts
存在命名冲突（ErrorHandler 接口与文件名混淆）
- 改了什么：将 ErrorHandler.ts 移动并重命名为 errors/ErrorHelper.ts，同时移动相关测试文件
- 影响范围：
  - 新增：apps/backend/errors/ErrorHelper.ts 及测试文件
  - 删除：apps/backend/services/ErrorHandler.ts 及测试文件
  - 修改：errors/index.ts（命名空间导出）、services/index.ts（移除导出）
  - 无业务代码依赖此模块，仅测试文件使用，无需向后兼容处理
- 验证方式：类型检查通过、567个测试全部通过、代码风格检查通过

**类型冲突解决方案：**
使用命名空间导出避免与 MCPErrors.ts 中的 ErrorCategory、MCPError 类型冲突：
```typescript
export * from "./MCPErrors.js";
export * as ErrorHelper from "./ErrorHelper.js";

文件变更清单：
- services/ErrorHandler.ts → errors/ErrorHelper.ts
- services/tests/ErrorHandler.test.ts → errors/tests/ErrorHelper.test.ts
- services/tests/AdvancedFeaturesIntegration.test.ts → errors/tests/ErrorHelper.integration.test.ts
