+++ apps/backend/errors/__tests__/ErrorHelper.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { Logger } from "../../Logger.js";
-import type { MCPError } from "../ErrorHandler.js";
+import type { MCPError } from "../ErrorHelper.js";
 import {
   ErrorCategory,
   RecoveryStrategy,
@@ -9,7 +9,7 @@ import {
   formatUserFriendlyMessage,
   getErrorStatistics,
   shouldAlert,
-} from "../ErrorHandler.js";
+} from "../ErrorHelper.js";
 
 // Mock dependencies
 vi.mock("../../Logger.js");
