+++ apps/backend/errors/index.ts
@@ -1,1 +1,4 @@
 export * from "./MCPErrors.js";
+
+// ErrorHelper 使用命名空间导出以避免与 MCPErrors 中的类型冲突
+export * as ErrorHelper from "./ErrorHelper.js";
