+++ apps/backend/services/index.ts
@@ -2,8 +2,5 @@ export * from "./StatusService.js";
 export * from "./NotificationService.js";
 export * from "./EventBus.js";
 
-// 新增导出 - 高优先级服务模块
-export * from "./ErrorHandler.js";
-
 // CustomMCPHandler 重新导出 - 保持向后兼容性
 export { CustomMCPHandler } from "@/lib/mcp/custom.js";
